# gohostdelivery-service-api-rest/src/main/resources/static/openapi.yml
openapi: 3.0.1
info:
  title: gohostdelivery API
  version: 1.0.0
  description: API para gestión de Riders, Companies, Documentos, Zonas de reparto y Pedidos en gohostdelivery.com
servers:
  - url: https://gohostdelivery-service-541341000706.europe-southwest1.run.app
    description: Preproducción (Cloud Run)
  - url: https://api.gohostdelivery.com
    description: Producción
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
paths:
...
  /riders:
    post:
      summary: Crear un nuevo Rider
      operationId: createRider
      tags: [Riders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderCreateRequest'
      responses:
        '201':
          description: Rider creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    get:
      summary: Listar Riders
      operationId: listRiders
      tags: [Riders]
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /riders/{riderId}:
    get:
      summary: Obtener Rider por ID
      operationId: getRiderById
      tags: [Riders]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rider encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    patch:
      summary: Actualización parcial de un Rider
      operationId: updateRiderPartial
      tags: [Riders]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderPatchRequest'
      responses:
        '200':
          description: Rider actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /riders/{riderId}/zones:
    get:
      summary: Obtener las zonas asignadas a un Rider
      operationId: getZonesByRider
      tags: [Riders]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de zonas asignadas al Rider
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Zone'
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }

  ###############################################################
  # RIDER DOCUMENTS (FOTO, CONTRATO, DOCUMENTO IDENTIDAD)
  ###############################################################
  /riders/{riderId}/documents:
    get:
      summary: Listar documentos de un Rider
      operationId: listRiderDocuments
      tags: [RiderDocuments]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de documentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RiderDocument'
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /riders/{riderId}/upload-photo:
    post:
      summary: Subir foto de perfil del Rider
      description: |
        Tamaño máximo permitido: **10 MB**.
        Formatos permitidos: JPG, PNG.
      operationId: uploadRiderPhoto
      tags: [RiderDocuments]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '201':
          description: Foto subida correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderDocument'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /riders/{riderId}/upload-contract:
    post:
      summary: Subir contrato del Rider
      description: |
        Tamaño máximo permitido: **20 MB**.
        Formatos permitidos: PDF, JPG, PNG.
      operationId: uploadRiderContract
      tags: [RiderDocuments]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '201':
          description: Contrato subido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderDocument'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /riders/{riderId}/upload-document:
    post:
      summary: Subir documento de identidad del Rider
      description: |
        Tamaño máximo permitido: **10 MB**.
        Formatos permitidos: PDF, JPG, PNG.
      operationId: uploadRiderIdentityDocument
      tags: [RiderDocuments]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '201':
          description: Documento subido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiderDocument'
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /riders/{riderId}/documents/{documentId}:
    delete:
      summary: Eliminar documento de un Rider
      operationId: deleteRiderDocument
      tags: [RiderDocuments]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204': { description: Documento eliminado correctamente }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /riders/{riderId}/documents/{documentId}/download:
    get:
      summary: Descargar documento de un Rider
      operationId: downloadRiderDocument
      tags: [RiderDocuments]
      parameters:
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Documento descargado correctamente
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }

  ###############################################################
  # COMPANIES
  ###############################################################
...
  /companies/{companyId}/riders/{riderId}:
    delete:
      summary: Eliminar preferencia de Rider de una Company
      operationId: removeRiderPreference
      tags: [Companies]
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204': { description: Preferencia eliminada correctamente }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }

  ###############################################################
  # ZONAS DE REPARTO
  ###############################################################
...
  /zones/{zoneId}/riders/{riderId}:
    post:
      summary: Asociar un Rider a una zona de reparto
      operationId: assignRiderToZone
      tags: [Zones]
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Rider asociado correctamente a la zona
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }

    delete:
      summary: Desasociar un Rider de una zona de reparto
      operationId: removeRiderFromZone
      tags: [Zones]
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: riderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Rider desasociado correctamente de la zona
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    OK:
      description: Successful operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SearchResponse'
    BadRequest:
      description: Bad request due to invalid parameters or malformed input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  ###############################################################
  # SCHEMAS
  ###############################################################
  schemas:
...
    Rider:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        apellido:
          type: string
        docIdentidad:
          type: string
        direccion:
          type: string
        municipio:
          type: string
        provincia:
          type: string
        telefono:
          type: string
        email:
          type: string
          format: email
        activo:
          type: boolean
      required:
        - id
        - nombre
        - apellido
        - docIdentidad
        - direccion
        - municipio
        - provincia
        - telefono
        - email
    RiderCreateRequest:
      type: object
      properties:
        nombre: { type: string }
        apellido: { type: string }
        docIdentidad: { type: string }
        direccion: { type: string }
        municipio: { type: string }
        provincia: { type: string }
        telefono: { type: string }
        email:
          type: string
          format: email
      required:
        - nombre
        - apellido
        - docIdentidad
        - direccion
        - municipio
        - provincia
        - telefono
        - email
...
    RiderPatchRequest:
      type: object
      properties:
        nombre: { type: string }
        apellido: { type: string }
        docIdentidad: { type: string }
        direccion: { type: string }
        municipio: { type: string }
        provincia: { type: string }
        telefono: { type: string }
        email: { type: string }
        activo: { type: boolean }
    RiderDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        riderId:
          type: string
          format: uuid
        tipoDocumento:
          type: string
          enum:
            - PROFILE_PHOTO
            - CONTRACT
            - ID_DOCUMENT
        nombreFichero:
          type: string
        contentType:
          type: string
        fechaSubida:
          type: string
          format: date-time
      required:
        - id
        - riderId
        - tipoDocumento
        - nombreFichero
        - contentType
        - fechaSubida

    ###############################################################
    # COMPANY
    ###############################################################
...
    Zone:
      type: object
      properties:
        idZone:
          type: string
          format: uuid
        idCompany:
          type: string
          format: uuid
        idCity:
          type: string
          format: uuid
        importeZonaCentro:
          type: number
          format: double
        importeZonaPeriferica:
          type: number
          format: double
        descripcion:
          type: string
          maxLength: 500
      required:
        - idZone
        - idCompany
        - idCity
        - importeZonaCentro
        - importeZonaPeriferica
...
    SearchResponse:
      type: object
      properties:
        total: { type: integer }
        items:
          type: array
          items:
            type: object
    ErrorResponse:
      type: object
      properties:
        errorCode: { type: string }
        message: { type: string }
      required:
        - errorCode
        - message